# be-summit/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BE Summit API Documentation",
        "description": "API documentation for the BE Summit project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://be-summit.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/gunung": {
            "post": {
                "tags": [
                    "Mountain (Admin)"
                ],
                "summary": "Create a new mountain profile (Admin)",
                "operationId": "1330865348b204cfdc490b4744879482",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_gunung",
                                    "deskripsi",
                                    "tinggi_mdpl",
                                    "lokasi",
                                    "foto",
                                    "status"
                                ],
                                "properties": {
                                    "nama_gunung": {
                                        "type": "string",
                                        "example": "Gunung Gede"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Gunung berapi aktif di Jawa Barat."
                                    },
                                    "tinggi_mdpl": {
                                        "type": "integer",
                                        "example": 2958
                                    },
                                    "lokasi": {
                                        "type": "string",
                                        "example": "Cianjur/Sukabumi/Bogor, Jawa Barat"
                                    },
                                    "foto": {
                                        "description": "Mountain picture file (JPEG/PNG, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "aktif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mountain created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mountain created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GunungResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gunung/{id}": {
            "post": {
                "tags": [
                    "Mountain (Admin)"
                ],
                "summary": "Update an existing mountain profile (Admin) [Uses POST with _method=PUT]",
                "operationId": "3723c05351d0181cc78dbe8a675b62e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mountain ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_gunung",
                                    "deskripsi",
                                    "tinggi_mdpl",
                                    "lokasi",
                                    "status"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "default": "PUT"
                                    },
                                    "nama_gunung": {
                                        "type": "string",
                                        "example": "Gunung Gede Baru"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Gunung berapi di Jawa Barat dengan pemandangan indah."
                                    },
                                    "tinggi_mdpl": {
                                        "type": "integer",
                                        "example": 2958
                                    },
                                    "lokasi": {
                                        "type": "string",
                                        "example": "Cianjur, Jawa Barat"
                                    },
                                    "foto": {
                                        "description": "Optional new mountain picture file (JPEG/PNG, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "aktif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mountain updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mountain updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/GunungResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Mountain not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mountain (Admin)"
                ],
                "summary": "Delete a mountain profile (Admin)",
                "operationId": "697e9581556635c9ae8dc2afac5835db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mountain ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mountain deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mountain deleted successfully."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Mountain not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GunungResource": {
                "title": "Gunung Resource",
                "description": "Mountain profile resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_gunung": {
                        "type": "string",
                        "example": "Gunung Gede"
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Gunung api aktif di Jawa Barat yang populer."
                    },
                    "tinggi_mdpl": {
                        "type": "integer",
                        "example": 2958
                    },
                    "lokasi": {
                        "type": "string",
                        "example": "Cianjur, Jawa Barat"
                    },
                    "foto": {
                        "type": "string",
                        "example": "http://be-summit.test/storage/gunungs/gede.jpg"
                    },
                    "status": {
                        "type": "string",
                        "example": "aktif"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-03T12:00:00+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-03T12:00:00+07:00"
                    },
                    "jalurs": {
                        "description": "List of trails associated with this mountain",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JalurPendakianResource"
                        }
                    }
                },
                "type": "object"
            },
            "JalurPendakianResource": {
                "title": "Jalur Pendakian Resource",
                "description": "Trail profile resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "gunung_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_jalur": {
                        "type": "string",
                        "example": "Jalur Cibodas"
                    },
                    "deskripsi": {
                        "type": "string",
                        "example": "Jalur legendaris berbatu dengan pemandangan air panas."
                    },
                    "titik_awal_mdpl": {
                        "type": "string",
                        "example": "1300 MDPL"
                    },
                    "titik_akhir_mdpl": {
                        "type": "string",
                        "example": "2958 MDPL"
                    },
                    "waktu_tempuh": {
                        "type": "string",
                        "example": "7 Jam"
                    },
                    "status": {
                        "type": "string",
                        "example": "open",
                        "enum": [
                            "open",
                            "close"
                        ]
                    },
                    "panjang_jalur": {
                        "type": "string",
                        "example": "9.7 Km"
                    },
                    "tingkat_kesulitan": {
                        "type": "string",
                        "example": "sedang",
                        "enum": [
                            "mudah",
                            "sedang",
                            "sulit",
                            "ekstrem"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-03T12:00:00+07:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-03T12:00:00+07:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}